removed unused protocols EncoderProtocol and DecoderProtocol
